\name{ENmisc-package}
\alias{ENmisc-package}
\alias{ENmisc}
\docType{package}
\title{Neuwirth Miscellaneous}
\description{
Utility functions: mtapply and mlapply (multivarate tapply and lapply) 
   wtd.boxplot (weighted boxplot), visual interface to generate mosaic plots
}
\details{
\tabular{ll}{
Package: \tab ENmisc\cr
Type: \tab Package\cr
License: \tab GPL 2.0\cr
LazyLoad: \tab yes\cr
}
}
\author{
Erich Neuwirth (mosaic tools developed after discussions with Rich Heiberger)

Maintainer: Erich Neuwirth <erich.neuwirth@univie.ac.at>
}

\keyword{ programming 
utilities 
hplot 
}

\note{Special thanks go to Rich Heiberger and John Verzani. 

Extended discussions with Rich Heiberger helped
bringing the mosaic part of the package in its final shape.
He invested quite some time in testing
uncharted territory and made valuable suggestions for improving
the interface to mosaic.

Without John Verzani's package \code{gWidgets}
and his always helpful comments the graphical
interface for mosaic plots would not exist.}
