\name{ENmisc-package}
\alias{ENmisc-package}
\alias{ENmisc}
\docType{package}
\title{Neuwirth Miscellaneous}
\description{
Utility functions, currently mtapply (multivarate tapply) and wtd.boxplot
(weighted boxplot)
}
\details{
\tabular{ll}{
Package: \tab ENmisc\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2010-12-29\cr
License: \tab GPL 2.0\cr
LazyLoad: \tab yes\cr
}
}
\author{
Erich Neuwirth

Maintainer: Erich Neuwirth <erich.neuwirth@univie.ac.at>
}
\keyword{ programming 
utilities 
hplot 
}
